Complete the home page and add responsive design for mobile devices
