+++ frontend/src/App.css
@@ -1,0 +1,17 @@
+.loader {
+    border: 13px solid #8ca59d;
+    border-top: 13px solid #f3f3f3;
+    border-radius: 50%;
+    width: 70px;
+    height: 70px;
+    animation: spin 1.5s linear infinite;
+}
+
+@keyframes spin {
+    0% {
+        transform: rotate(0deg);
+    }
+    100% {
+        transform: rotate(360deg);
+    }
+}
