+++ frontend/src/pages/dashboards/admin/adminOverview.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const AdminOverview = () => {
+  return (
+    <div>AdminOverview</div>
+  )
+}
+
+export default AdminOverview
