+++ frontend/src/pages/dashboards/admin/manageReservations.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ManageReservations = () => {
+  return (
+    <div>ManageReservations</div>
+  )
+}
+
+export default ManageReservations
