+++ frontend/src/pages/dashboards/admin/manageSeats.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ManageSeats = () => {
+  return (
+    <div>ManageSeats</div>
+  )
+}
+
+export default ManageSeats
