+++ frontend/src/pages/dashboards/admin/reports.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Reports = () => {
+  return (
+    <div>Reports</div>
+  )
+}
+
+export default Reports
